moved cache and simplified code and removed unused things
