+++ nativeApp/src/lib/ui/modals/InvitationModal.tsx
@@ -3,7 +3,6 @@ import API from '../../../api';
 import styled from 'styled-components/native';
 import Modal from 'react-native-modal';
 import { Button as B, Card as C, Textarea as TA, Form } from 'native-base';
-import { Image } from 'react-native';
 import { AddNote, Close } from '../AddNote';
 
 import { Heading2 as H2, FontsEnum } from '../typography';
@@ -16,28 +15,19 @@ type Props = {
   isVisible: boolean;
   onSendClose: () => void;
   onCancel: () => void;
-  awardBadge: (awardBadgeState) => boolean;
+  awardBadge: (awardBadgeState) => void;
+  title: string;
 }
 
 /*
  * Styles
  */
-const View = styled.View`
-`;
-
 const Heading2 = styled(H2)`
   paddingHorizontal: 15;
   paddingTop: 15;
   marginBottom: 20;
   color: ${ColoursEnum.white};
 `;
-
-// const Card = styled(C)`
-//   paddingHorizontal: 15;
-//   paddingTop: 15;
-//   paddingBottom: 10;
-//   borderRadius: 5;
-// `;
 
 const Card = styled(C)`
   paddingBottom: 10;
